+++ qtbook_book.cc
@@ -543,8 +543,8 @@ void qtbook_book::slotGo(void)
 	query.bindValue(22, oid);
       else if(qmain->getDB().driverName() == "QSQLITE")
 	{
-	  int value = misc_functions::getSqliteUniqueId(qmain->getDB(),
-							errorstr);
+	  qint64 value = misc_functions::getSqliteUniqueId(qmain->getDB(),
+							   errorstr);
 
 	  if(errorstr.isEmpty())
 	    query.bindValue(22, value);
